pyproject.toml
[project]
name = "coral"
version = "0.1.0"
description = "Coral is an AI agent, originally from TitanOS Nebula. It has migrated into a standalone agent, including integrations with other platforms."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ddgs>=9.11.4",
    "discord-py>=2.7.1",
    "jinja2>=3.1.6",
    "pydantic>=2.12.5",
    "pydantic-ai>=1.70.0",
    "pyyaml>=6.0.3",
    "sqlmodel>=0.0.37",
    "typer>=0.24.1",
]

[project.scripts]
coral = "coral.main:app"
